cmake

+++ deck.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 
 namespace Canasta {
-Deck::Deck() {
-    int size = this->cards.size();
-    std::cout << "constructor called " << size << std::endl;
-}
+    Deck::Deck() {
+        size_t size = this->cards.size();
+        std::cout << "constructor called " << size << std::endl;
+    }
 };
